【青岛】 兑吧订单 订单列表导出 售价字段更改为INT类型

+++ hyjf_platform_server/hyjf_admin/src/main/java/com/hyjf/admin/controller/pointsshop/duiba/order/DuibaOrderController.java
@@ -327,7 +327,7 @@ public class DuibaOrderController extends BaseController {
             @Override
             public String format(Object object) {
                 BigDecimal sellingPrice = (BigDecimal) object;
-                return sellingPrice + "";
+                return sellingPrice.intValue() + "";
             }
         };
         IValueFormatter markingPriceAdapter = new IValueFormatter() {
